requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu \
torch == 2.1.2 \
torchaudio == 2.1.2

pyTelegramBotAPI
openai-whisper